requirements.txt
pyserial
keyboard
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt==4.0.1
python-jose
python-multipart
requests